Added modified ball test
This is based on Will's bouncing ball test. This version was made so I could familiarize myself with working with SDL outside of the LazyFoo tutorials. This version has less bounciness, a debug.txt log, and some code to stop a physics bug caused by moving the game window around.